week07/docker-pipeline/etl_job/requirements.txt
python-dotenv
pymongo
SQLAlchemy
psycopg2-binary